requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-dotenv==1.0.0
Flask-Login==0.6.3
SQLAlchemy==2.0.15
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
